SQUASH: Try update endpoint URL
The URL doesn't need to mention the action (update),
that is implied via the HTTP method (POST).

Instead, we can abstract this a bit and use it again in the future.
The other fork we had of `get_roster` could be expressed as a GET of
this same URL here (get the enrollment status of a course).